refactor: added containers and components folder to differ from pages
