+++ components/card-with-photo.tsx
@@ -1,3 +1,5 @@
+import React from 'react';
+
 export const CardWithPhoto = () => {
   return (
     <div className="card">
