+++ pages/pets/index.tsx
@@ -1,11 +1,7 @@
 import React from 'react';
-import { PetsListContainer } from './pets-list.container';
+import { PetsListContainer } from '../../containers/pets/pets-list';
 
 const PetsPage = () => {
-  // const {data, error} = useSWR('/api/pets', getPets);
-  // console.log(data, error);
-  // if (error) return <div>Erro ao carregar os pets</div>
-  // if (!data) return <div>Carregando os pets</div>
   return (
     <div className="container">
       <div className="has-text-centered" style={{ marginTop: '5%' }}>
